feat: criteria

+++ lib/data/model/request_add_story.dart
@@ -4,8 +4,8 @@ import 'dart:io';
 class AddStoryRequest {
   String description;
   File photo;
-  Float? lat;
-  Float? lon;
+  double? lat;
+  double? lon;
 
   AddStoryRequest(
       {required this.description, required this.photo, this.lat, this.lon});
